# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Rule Validation API",
    "version": "1.0.0",
    "description": "API for managing and validating rules."
  },
  "paths": {
    "/rules": {
      "get": {
        "summary": "Get all rules",
        "description": "Retrieve a list of all rules. Can filter by is_active.",
        "parameters": [
          {
            "name": "is_active",
            "in": "query",
            "description": "Filter rules by active status (true/false)",
            "required": false,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "A list of rules."
          }
        }
      },
      "post": {
        "summary": "Create a new rule",
        "description": "Create a new rule with the provided details.",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "rule",
            "description": "The rule to create.",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "description": "The name of the rule."
                },
                "description": {
                  "type": "string",
                  "description": "A description of the rule."
                },
                "condition": {
                  "type": "string",
                  "description": "The condition for the rule."
                },
                "action": {
                  "type": "string",
                  "description": "The action to take when the condition is met."
                }
              },
              "required": [
                "name",
                "description",
                "condition",
                "action"
              ]
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Rule created successfully."
          },
          "400": {
            "description": "Invalid input."
          }
        }
      }
    },
    "/rules/{id}": {
      "get": {
        "summary": "Get a rule by ID",
        "description": "Retrieve a specific rule by its ID.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the rule to retrieve.",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested rule."
          },
          "404": {
            "description": "Rule not found."
          }
        }
      },
      "put": {
        "summary": "Update a rule by ID",
        "description": "Update an existing rule with the provided details.",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the rule to update.",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "rule",
            "description": "The updated rule details.",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "description": "The name of the rule."
                },
                "description": {
                  "type": "string",
                  "description": "A description of the rule."
                },
                "condition": {
                  "type": "string",
                  "description": "The condition for the rule."
                },
                "action": {
                  "type": "string",
                  "description": "The action to take when the condition is met."
                },
                "is_active": {
                  "type": "boolean",
                  "description": "Whether the rule is active."
                }
              },
              "required": [
                "name",
                "description",
                "condition",
                "action"
              ]
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Rule updated successfully."
          },
          "400": {
            "description": "Invalid input."
          },
          "404": {
            "description": "Rule not found."
          }
        }
      },
      "delete": {
        "summary": "Delete a rule by ID",
        "description": "Delete a specific rule by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the rule to delete.",
            "required": true,
            "type": "integer"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Rule deleted successfully."
          },
          "404": {
            "description": "Rule not found."
          }
        }
      }
    }
  }
}